+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+See [fix-counts.js](scripts/updates/fix-count.js) for a script that can be run before this update (as well as the
+update for 1.19.0) to pre populate the migration. This will speed up the update and will not impact the running
+instance.
+
+### Fixed
+- If a space has a lot of datasets, rendering the space page is very slow. Files in a space is now cached.
+
 ## 1.19.2 - 2021-10-20
 
 ### Fixed